♻ Refactored frequency to signals
